pyproject.toml
[tool.poetry]
name = "langchain-note-2024"
version = "0.1.0"
description = ""
authors = ["wemade_june <nobody.jbj@wemade.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
langchain = "^0.3.7"
openapi = "^2.0.0"
langchain-openai = "^0.2.6"
langsmith = "^0.1.140"
langchain-community = "^0.3.5"
langchain-chroma = "^0.1.4"
streamlit = "^1.40.1"
pymupdf = "^1.24.13"
faiss-cpu = "^1.9.0"
pydantic = "^2.9.2"
google-search-results = "^2.4.2"
langchain-google-genai = "^2.0.5"
langchain-experimental = "^0.3.3"
feedparser = "^6.0.11"
bs4 = "^0.0.2"
matplotlib = "^3.9.3"
pypdf = "^5.1.0"
panadas = "^0.2"
tabulate = "^0.9.0"
seaborn = "^0.13.2"
mplfinance = "^0.12.10b0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
